Initial Upload
Contains accelerometer plugin and Dialogue System for Unity with CYOA
Framework add in. I added the default .gitignore for Unity from the list
on GitHub to minimize the upload, but you might need the whole thing. I
will zip the entire project and upload to Google Drive and send you a
URL.